Added skip for not-i18n doc dirs to translation status update script.

+++ src/TRANSLATION_STATUS_UPDATE.sh
@@ -13,12 +13,15 @@ export LC_ALL=C
 
 for i in $DIRLIST
 do
-  for j in $i/po/*.po
-  do
-    NNLANG=`basename $j .po`
-    echo -n "$i, $NNLANG, " >>$OUTPUTF
-    msgfmt --use-fuzzy --statistics $j 2>&1 |sed  -e 's/,/;/g' -e 's/\.//' -e 's/ messages//g'  -e 's/ message//g' -e 's/ translation//' -e 's/ translations//'  >> $OUTPUTF
-    #  
-    rm messages.mo
-  done
+  #check for i18n
+  if [ -e "$i/po" ]; then
+    for j in $i/po/*.po
+    do
+      NNLANG=`basename $j .po`
+      echo -n "$i, $NNLANG, " >>$OUTPUTF
+      msgfmt --use-fuzzy --statistics $j 2>&1 |sed  -e 's/,/;/g' -e 's/\.//' -e 's/ messages//g'  -e 's/ message//g' -e 's/ translation//' -e 's/ translations//'  >> $OUTPUTF
+      #  
+      rm messages.mo
+    done
+  fi
 done
